created simple panel group
