+++ src/layout/Group/GroupRenderer.tsx
@@ -5,6 +5,7 @@ import { DisplayGroupContainer } from 'src/layout/Group/DisplayGroupContainer';
 import { GroupContainer } from 'src/layout/Group/GroupContainer';
 import { RepeatingGroupsFocusProvider } from 'src/layout/Group/RepeatingGroupsFocusContext';
 import { PanelGroupContainer } from 'src/layout/Panel/PanelGroupContainer';
+import { PanelReferenceGroupContainer } from 'src/layout/Panel/PanelReferenceGroupContainer';
 import type { PropsFromGenericComponent } from 'src/layout';
 
 export type GroupRendererProps = PropsFromGenericComponent<'Group'>;
@@ -22,11 +23,22 @@ export function GroupRenderer({ node }: GroupRendererProps) {
     );
   }
 
+  // panel with groupReference
+  if (node.item.panel?.groupReference) {
+    return (
+      <PanelReferenceGroupContainer
+        key={node.item.id}
+        id={node.item.id}
+      />
+    );
+  }
+
+  // regular panel group
   if (node.item.panel) {
     return (
       <PanelGroupContainer
         key={node.item.id}
-        id={node.item.id}
+        node={node}
       />
     );
   }
