+++ src/layout/Panel/PanelGroupContainer.module.css
@@ -1,0 +1,3 @@
+.panelPadding > div {
+  --panel-x-padding: var(--modal-padding-x) !important;
+}
